Added Elevator Prefab
Provided Prefab is an elevator, that is triggered by a keypress.
Provide it an initial state, a movement speed, and a time to move.
Only movement in the +/- y direction is currently possible.  The
initial state should only be STATIONARY_TOP or STATIONARY_BOTTOM.
Characters on a moving elevator won’t move.