Implemented Pause Functionality, and changed CharacterMove collsision detection slightly
Added a “Pause Watcher” prefab.  If added to a level, the player can
pause at any time by pressing the pause button, which is “p” by default
(and can be changed through the PauseListener component of the prefab).

Added to CharacterMove.FixedUpdate the following test: If it has not
moved since the previous FixedUpdate, turn around.  This was added to
fix a bug that would cause two characters to run into each other and
then stop.